Move run_program test helper into runner module
