+++ lib/src/metta/runner/stdlib/debug.rs
@@ -329,7 +329,7 @@ mod tests {
     use super::*;
     use crate::metta::runner::{Metta, EnvBuilder, SExprParser};
     use crate::common::test_utils::metta_space;
-    use crate::metta::runner::stdlib::tests::run_program;
+    use crate::metta::runner::run_program;
 
     use regex::Regex;
